renamed each feature modules
