+++ feature-twitch/src/main/AndroidManifest.xml
