+++ feature-twitch/src/main/java/com/freshdigitable/yttt/feature/timetable/TimetableContextMenuDelegateForTwitch.kt
@@ -1,9 +1,7 @@
-package com.freshdigitable.yttt.feature.timetable.twitch
+package com.freshdigitable.yttt.feature.timetable
 
 import com.freshdigitable.yttt.LaunchAppWithUrlUseCase
 import com.freshdigitable.yttt.data.model.LiveVideo
-import com.freshdigitable.yttt.feature.timetable.TimetableContextMenuSelector
-import com.freshdigitable.yttt.feature.timetable.TimetableMenuItem
 import javax.inject.Inject
 
 internal class TimetableContextMenuDelegateForTwitch @Inject constructor(
