+++ feature-youtube/src/main/AndroidManifest.xml
