+++ feature-youtube/src/main/java/com/freshdigitable/yttt/feature/timetable/FetchYouTubeStreamUseCase.kt
@@ -1,4 +1,4 @@
-package com.freshdigitable.yttt.feature.timetable.youtube
+package com.freshdigitable.yttt.feature.timetable
 
 import android.util.Log
 import com.freshdigitable.yttt.data.AccountRepository
@@ -8,7 +8,6 @@ import com.freshdigitable.yttt.data.YouTubeRepository
 import com.freshdigitable.yttt.data.model.YouTubeSubscriptionSummary
 import com.freshdigitable.yttt.data.model.YouTubeSubscriptionSummary.Companion.needsUpdatePlaylist
 import com.freshdigitable.yttt.data.model.YouTubeVideo
-import com.freshdigitable.yttt.feature.timetable.FetchStreamUseCase
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
