+++ feature-youtube/src/main/java/com/freshdigitable/yttt/feature/timetable/TimetableContextMenuDelegateForYouTube.kt
@@ -1,11 +1,9 @@
-package com.freshdigitable.yttt.feature.timetable.youtube
+package com.freshdigitable.yttt.feature.timetable
 
 import com.freshdigitable.yttt.LaunchAppWithUrlUseCase
 import com.freshdigitable.yttt.data.YouTubeRepository
 import com.freshdigitable.yttt.data.model.LiveVideo
 import com.freshdigitable.yttt.data.model.mapTo
-import com.freshdigitable.yttt.feature.timetable.TimetableContextMenuSelector
-import com.freshdigitable.yttt.feature.timetable.TimetableMenuItem
 import javax.inject.Inject
 
 internal class TimetableContextMenuDelegateForYouTube @Inject constructor(
