+++ feature-youtube/src/main/java/com/freshdigitable/yttt/feature/timetable/di/YouTubeTimetableFeatureModule.kt
@@ -1,4 +1,4 @@
-package com.freshdigitable.yttt.feature.timetable.youtube.di
+package com.freshdigitable.yttt.feature.timetable.di
 
 import com.freshdigitable.yttt.data.model.YouTubeVideo
 import com.freshdigitable.yttt.di.IdBaseClassKey
@@ -7,11 +7,11 @@ import com.freshdigitable.yttt.feature.timetable.FetchTimetableItemSourceUseCase
 import com.freshdigitable.yttt.feature.timetable.TimetableContextMenuSelector
 import com.freshdigitable.yttt.feature.timetable.TimetablePage
 import com.freshdigitable.yttt.feature.timetable.TimetableTabQualifier
-import com.freshdigitable.yttt.feature.timetable.youtube.FetchYouTubeFreeChatItemSourceUseCase
-import com.freshdigitable.yttt.feature.timetable.youtube.FetchYouTubeOnAirItemSourceUseCase
-import com.freshdigitable.yttt.feature.timetable.youtube.FetchYouTubeStreamUseCase
-import com.freshdigitable.yttt.feature.timetable.youtube.FetchYouTubeUpcomingItemSourceUseCase
-import com.freshdigitable.yttt.feature.timetable.youtube.TimetableContextMenuDelegateForYouTube
+import com.freshdigitable.yttt.feature.timetable.FetchYouTubeFreeChatItemSourceUseCase
+import com.freshdigitable.yttt.feature.timetable.FetchYouTubeOnAirItemSourceUseCase
+import com.freshdigitable.yttt.feature.timetable.FetchYouTubeStreamUseCase
+import com.freshdigitable.yttt.feature.timetable.FetchYouTubeUpcomingItemSourceUseCase
+import com.freshdigitable.yttt.feature.timetable.TimetableContextMenuDelegateForYouTube
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
