+++ settings.gradle
@@ -16,8 +16,8 @@ rootProject.name = "yttt"
 include ':app'
 include ':lib'
 include ':common'
-include ':feature-timetable-youtube'
-include ':feature-timetable-twitch'
+include ':feature-youtube'
+include ':feature-twitch'
 include ':repository-youtube'
 include ':repository-twitch'
 include ':repository-appuser'
